Implement authorization using Gate

+++ app/Http/Controllers/QuestionsController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\AskQuestionRequest;
 use App\Question;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionsController extends Controller
 {
@@ -71,6 +72,11 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(Gate::denies('update-question', $question))
+        {
+            abort(403, "Access denied");
+        }
+
         return view('questions.edit', compact('question'));
     }
 
@@ -83,6 +89,11 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(Gate::denies('update-question', $question))
+        {
+            abort(403, "Access denied");
+        }
+
         $question->update($request->only('title','body'));
 
         return redirect('/questions')->with('success','Your Question has been updated.');
@@ -96,6 +107,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(Gate::denies('delete-question', $question))
+        {
+            abort(403, "Access denied");
+        }
         //dd($question);
         $question->delete();
